+++ weed/log_buffer/log_buffer.go
@@ -1,4 +1,4 @@
-package queue
+package log_buffer
 
 import (
 	"sync"
